[chore #156406804] Add Model helper methods
